+++ packages/domain-clients-core/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Signing and public clients for solana
+- Config constructor for solana
 
 ## [0.2.0] 2025-08-19